ARM: OMAP: hwmod: minor fix for return value of _deassert_hardreset
The return value is declared as int to be in line with the
function definition.

+++ arch/arm/mach-omap2/omap_hwmod.c
@@ -1579,7 +1579,7 @@ static int _deassert_hardreset(struct omap_hwmod *oh, const char *name)
 static int _read_hardreset(struct omap_hwmod *oh, const char *name)
 {
 	struct omap_hwmod_rst_info ohri;
-	u8 ret = -EINVAL;
+	int ret = -EINVAL;
 
 	if (!oh)
 		return -EINVAL;
